Add data layer on sample (WIP) - Network layer basic implementation

+++ sample/src/main/java/com/mobilejazz/kotlin/core/sample/domain/model/Item.kt
@@ -2,10 +2,9 @@ package com.mobilejazz.kotlin.core.sample.domain.model
 
 import java.math.BigDecimal
 
-
 data class Item(val id: String?,
                 val name: String,
-                val price: BigDecimal,
+                val price: Double,
                 val count: Int,
                 val imageUrl: String){
